refactor: split json_parser.py into focused sub-modules (_report_parsing, _principle_map, _eval_parsing)
Each sub-module is under 100 lines. json_parser.py remains as re-export
facade for backward compatibility.
